refactor: reimport http api methods from api/http
